+++ src/Installer.ts
@@ -8,6 +8,7 @@ import { existsSync, rmSync } from "fs";
 import type { Language } from "tree-sitter";
 import { getLogger } from "./outputChannel";
 import { mkdir } from "fs/promises";
+import { pathToFileURL } from "node:url";
 import which from "which";
 
 const NPM_INSTALL_URL = "https://nodejs.org/en/download";
@@ -16,7 +17,7 @@ export function getAbsoluteParserDir(parsersDir: string, parserName: string): st
     return path.resolve(path.join(parsersDir, parserName));
 }
 
-export function getAbsoluteBindingsDir(parsersDir: string, parserName: string): string {
+export function getAbsoluteBindingsPath(parsersDir: string, parserName: string): string {
     return path.resolve(path.join(parsersDir, parserName, "bindings", "node", "index.js"));
 }
 
@@ -27,24 +28,17 @@ export async function loadParser(
 ): Promise<Result<Language, string>> {
     const logger = getLogger();
 
-    const bindingsDir = getAbsoluteBindingsDir(parsersDir, parserName);
-    if (!existsSync(bindingsDir)) {
-        const msg = `Expected parser directory doesn't exist: ${bindingsDir}`;
+    const bindingsPath = getAbsoluteBindingsPath(parsersDir, parserName);
+    if (!existsSync(bindingsPath)) {
+        const msg = `Expected parser bindings don't exist: ${bindingsPath}`;
         logger.log(msg);
         return err(msg);
     }
 
     try {
-        logger.log(`Loading parser from ${bindingsDir}`);
+        logger.log(`Loading parser from ${bindingsPath}`);
 
-        let language: Language;
-        try {
-            language = ((await import(bindingsDir)) as { default: Language }).default;
-        } catch (_) {
-            // TODO(1/11/25): Always use import and remove this backup
-            // eslint-disable-next-line @typescript-eslint/no-require-imports
-            language = require(bindingsDir) as Language;
-        }
+        let language = ((await import(pathToFileURL(bindingsPath).href)) as { default: Language }).default;
 
         logger.log(`Got language: ${JSON.stringify(Object.keys(language))}`);
 
@@ -58,8 +52,8 @@ export async function loadParser(
 
         return ok(language);
     } catch (error) {
-        logger.log(`Failed to load ${bindingsDir} > ${JSON.stringify(error)}`);
-        return err(`Failed to load ${bindingsDir} > ${JSON.stringify(error)}`);
+        logger.log(`Failed to load ${bindingsPath} > ${JSON.stringify(error)}`);
+        return err(`Failed to load ${bindingsPath} > ${JSON.stringify(error)}`);
     }
 }
 
@@ -68,7 +62,7 @@ export async function downloadAndBuildParser(
     parserNpmPackage: string,
     parserName: string,
     npm: string,
-    treeSitterCli: string,
+    rebuild: boolean = false,
     onData?: (data: string) => void
 ): Promise<Result<void, string>> {
     const logger = getLogger();
@@ -117,31 +111,37 @@ export async function downloadAndBuildParser(
     }
 
     // try to load parser optimistically
-    const loadResult = await loadParser(parsersDir, parserName);
-    if (loadResult.status === "ok") {
-        return ok(undefined);
+    if (!rebuild) {
+        const loadResult = await loadParser(parsersDir, parserName);
+        if (loadResult.status === "ok") {
+            return ok(undefined);
+        }
+        logger.log(`Optimistic load failed for parser ${parserName}`);
     }
 
-    logger.log(`Optimistic load failed, trying to build parser ${parserName}`);
-    const treeSitterCliOk = await runCmd(`${treeSitterCli} --version`);
-    if (treeSitterCliOk.status === "err") {
+    // if it fails, try to build it
+    logger.log(`Building parser ${parserName}`);
+    const installDepsResult = await runCmd(`${npm} install --loglevel=silly`, { cwd: parserDir }, (d) =>
+        onData?.(d.toString())
+    );
+    if (installDepsResult.status === "err") {
         const msg =
-            `Parser ${parserName} requires local build, but
-            tree-sitter cli command '${treeSitterCli}' failed:
-            ${treeSitterCliOk.result[0].name} ${treeSitterCliOk.result[0].message.replace(/\n/g, " > ")}.` +
-            (treeSitterCliOk.result[1].length > 1 ? ` Logs: ${treeSitterCliOk.result[1].join(">")}` : "");
+            "Failed to install parser dependencies > " +
+            installDepsResult.result[0].name +
+            ": " +
+            installDepsResult.result[0].message.replace(/\n/g, " > ") +
+            (installDepsResult.result[1].length > 1 ? ` Logs: ${installDepsResult.result[1].join(">")}` : "");
 
         logger.log(msg);
         return err(msg);
     }
 
-    // if it fails, try to build it
-    const buildResult = await runCmd(`${treeSitterCli} build`, { cwd: parserDir }, (d) =>
+    const buildResult = await runCmd(`${npm} rebuild --loglevel=silly`, { cwd: parserDir }, (d) =>
         onData?.(d.toString())
     );
     if (buildResult.status === "err") {
         const msg =
-            "Failed to build parser using tree-sitter cli > " +
+            "Failed to build parser > " +
             buildResult.result[0].name +
             ": " +
             buildResult.result[0].message.replace(/\n/g, " > ") +
@@ -191,7 +191,8 @@ export type GetLanguageError = {
 export async function getLanguage(
     parsersDir: string,
     languageId: string,
-    autoInstall = false
+    autoInstall = false,
+    rebuild = false
 ): Promise<Result<Language | undefined, GetLanguageError>> {
     const logger = getLogger();
 
@@ -205,7 +206,6 @@ export async function getLanguage(
     const parserPackagePath = getAbsoluteParserDir(parsersDir, parserName);
 
     const npm = "npm";
-    const treeSitterCli = configuration.getTreeSitterCliPath();
 
     if (!existsSync(parserPackagePath)) {
         const doInstall = autoInstall
@@ -258,7 +258,7 @@ export async function getLanguage(
                     npmPackageName,
                     parserName,
                     npm,
-                    treeSitterCli,
+                    rebuild,
                     (data) => progress.report({ message: data, increment: number++ })
                 );
             }